# Optimize rubocop