@koic 2022-12-07
Nice optimizing! 👍